refactor testdata for recovercheck
