+++ pkg/golinters/recovercheck/testdata/go.mod
@@ -1,0 +1,7 @@
+module testdata
+
+go 1.25
+
+require testdata/pkg v0.0.0-00010101000000-000000000000
+
+replace testdata/pkg => ./pkg
