+++ pkg/golinters/recovercheck/testdata/pkg/go.mod
@@ -1,0 +1,3 @@
+module testdata/pkg
+
+go 1.25
